+++ buildSrc/src/test/kotlin/com/example/model/Demo.kt
@@ -4,14 +4,14 @@ import com.google.gson.annotations.SerializedName
 import kotlin.Any
 import kotlin.Boolean
 import kotlin.Double
-import kotlin.Int
+import kotlin.Long
 import kotlin.String
 
 data class Demo(
     @SerializedName("s")
     val s: String,
     @SerializedName("i")
-    val i: Int,
+    val i: Long,
     @SerializedName("n")
     val n: Double,
     @SerializedName("b")
