+++ buildSrc/src/test/kotlin/com/example/model/Foo.kt
@@ -1,12 +1,12 @@
 package com.example.model
 
 import com.google.gson.annotations.SerializedName
-import kotlin.Int
+import kotlin.Long
 import kotlin.String
 
 data class Foo(
     @SerializedName("bar")
     val bar: String?,
     @SerializedName("baz")
-    val baz: Int?
+    val baz: Long?
 )
