+++ buildSrc/src/test/kotlin/com/example/model/Product.kt
@@ -2,12 +2,12 @@ package com.example.model
 
 import com.google.gson.annotations.SerializedName
 import kotlin.Double
-import kotlin.Int
+import kotlin.Long
 import kotlin.String
 
 data class Product(
     @SerializedName("productId")
-    val productId: Int,
+    val productId: Long,
     @SerializedName("productName")
     val productName: String,
     @SerializedName("price")
